requirements.txt
py-clob-client>=0.34.6
requests>=2.32.0
anthropic>=0.40.0
numpy>=1.26.0
tabulate>=0.9.0
python-dotenv>=1.0.0
aiohttp>=3.10.0
web3>=7.0.0
pydantic>=2.0.0
